multipathd: Ignore errors when creating pidfile
We can use CLI commands to communicate with the daemon,
so we don't need the pidfile for correct operation.
Hence any errors from creating the pidfile can be safely
ignored.

References: bnc#705546

+++ multipathd/main.c
@@ -1501,7 +1501,7 @@ child (void * param)
 	pthread_attr_t log_attr, misc_attr;
 	struct vectors * vecs;
 	sigset_t set;
-	int rc;
+	int rc, pid_rc;
 
 	mlockall(MCL_CURRENT | MCL_FUTURE);
 
@@ -1614,12 +1614,9 @@ child (void * param)
 	pthread_attr_destroy(&misc_attr);
 
 	/* Startup complete, create logfile */
-	if (pidfile_create(DEFAULT_PIDFILE, daemon_pid)) {
-		if (logsink)
-			log_thread_stop();
+	pid_rc = pidfile_create(DEFAULT_PIDFILE, daemon_pid);
+	/* Ignore errors, we can live without */
 
-		exit(1);
-	}
 	running_state = DAEMON_RUNNING;
 	pthread_cond_wait(&exit_cond, &exit_mutex);
 	/* Need to block these to avoid deadlocking */
@@ -1668,8 +1665,10 @@ child (void * param)
 	dm_lib_exit();
 
 	/* We're done here */
-	condlog(3, "unlink pidfile");
-	unlink(DEFAULT_PIDFILE);
+	if (!pid_rc) {
+		condlog(3, "unlink pidfile");
+		unlink(DEFAULT_PIDFILE);
+	}
 
 	condlog(2, "--------shut down-------");
